refactor(global): changed the components structure of the client to organize better
