+++ client/src/components/Dashboard/GroupsList/InstrumentList/InstrumentList.js
