+++ client/src/components/Dashboard/GroupsList/InstrumentList/index.js
