+++ client/src/components/Dashboard/ListParticipants/ParticipantItem/ParticipantItem.js
