+++ client/src/components/Dashboard/ListParticipants/ParticipantItem/index.js
