+++ client/src/components/Dashboard/ListParticipants/ParticipantList/index.js
