+++ client/src/components/ParticipantDetails/DetailsFormPayments/PaymentItem/index.js
@@ -1,0 +1,2 @@
+import PaymentItem from './PaymentItem';
+export default PaymentItem;
