--- client/src/components/PaymentDetail/index.js
@@ -1,2 +1,0 @@
-import PaymentDetail from './PaymentDetail';
-export default PaymentDetail;
