+++ client/src/components/Resources/Auth/auth0-provider-with-history.js
