+++ client/src/components/Resources/DeleteParticipantButton/DeleteParticipantButton.css
