+++ client/src/components/Resources/DeleteParticipantButton/DeleteParticipantButton.js
