+++ client/src/components/Resources/DeleteParticipantButton/index.js
