+++ client/src/components/Resources/Loading/Loading.js
