+++ client/src/components/Resources/Loading/index.js
