+++ client/src/components/Resources/Navbar/logout-button.js
