+++ client/src/components/Resources/SmartLink/SmartLink.js
