+++ client/src/components/Resources/SmartLink/index.js
