+++ client/src/components/Resources/StatusLight/StatusLight.js
